requirements/haystack-1.11.txt
absl-py
aiocontextvars==0.2.2
aiofiles==22.1.0
aiohttp
aiosignal
alembic==1.8.1
anyio==3.6.2
astunparse==1.6.3
async-timeout
attrs
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.1
backoff==1.11.1
blinker==1.4
blis==0.7.9
brotlipy==0.7.0
cachetools
catalogue==2.0.8
certifi
cffi
Chameleon==3.10.1
charset-normalizer
click
cloudpickle==2.2.0
confection==0.0.3
contextvars==2.4
contourpy==1.0.6
cryptography
cycler==0.11.0
cymem==2.0.7
databricks-cli==0.17.3
dill==0.3.6
docker==6.0.1
docopt==0.6.2
elasticsearch==7.17.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.4.1/en_core_web_trf-3.4.1-py3-none-any.whl
entrypoints==0.4
farm-haystack==1.11.0
fastapi==0.88.0
fastapi-chameleon==0.1.12
filelock==3.8.1
Flask==2.2.2
flatbuffers
flit_core
fonttools==4.38.0
frozenlist
gast
gitdb==4.0.10
GitPython==3.1.29
google-auth
google-auth-oauthlib
google-pasta
greenlet==2.0.1
grpcio==1.47.2
grpcio-tools==1.47.2
gunicorn==20.1.0
h11==0.14.0
h5py
huggingface-hub==0.11.1
idna
immutables==0.19
importlib-metadata
inflect==6.0.2
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.3
Jinja2==3.1.2
joblib==1.2.0
jq==1.3.0
jsonschema==4.17.3
keras
Keras-Preprocessing
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
libclang==14.0.6
llvmlite==0.39.1
loguru==0.6.0
lxml==4.9.1
Mako==1.2.4
Markdown
MarkupSafe==2.1.1
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
mlflow==2.0.1
mmh3==3.0.0
monotonic==1.6
more-itertools==9.0.0
msrest==0.7.1
multidict
murmurhash==1.0.9
networkx==2.8.8
nltk==3.7
num2words==0.5.12
numba==0.56.4
numpy
oauthlib
opt-einsum
packaging
pandas==1.5.2
pathy==0.10.0
Pillow==9.3.0
posthog==2.2.0
preshed==3.0.8
protobuf==3.19.6
pyarrow==10.0.1
pyasn1
pyasn1-modules==0.2.8
pycparser
pydantic==1.10.2
pygraphviz==1.10
PyJWT
pymilvus==2.2.0
pyOpenSSL
pyparsing
pyrsistent==0.19.2
PySocks
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.6
PyYAML==6.0
pyyaml-include==1.3
quantulum3==0.7.11
querystring-parser==1.2.4
rapidfuzz==2.7.0
regex==2022.10.31
requests
requests-oauthlib==1.3.0
rsa
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
seqeval==1.2.2
shap==0.41.0
six
slicer==0.0.7
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.0
spacy==3.4.3
spacy-alignments==0.8.6
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacy-transformers==1.1.8
SQLAlchemy==1.4.44
sqlparse==0.4.3
srsly==2.4.5
starlette==0.22.0
tabulate==0.9.0
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-hub==0.8.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor
thinc==8.1.5
threadpoolctl==3.1.0
tika==1.24
tokenizers==0.12.1
torch==1.12.1
# torchvision==0.13.1
tqdm==4.64.1
transformers==4.21.2
typer==0.7.0
typing_extensions
ujson==5.4.0
urllib3
uvicorn==0.20.0
venusian==3.0.0
wasabi==0.10.1
websocket-client==1.4.2
Werkzeug==2.2.2
wrapt
yarl
zipp
